Add mixins for each pretrained model; fix CI
- Add Pretrained*Mixin classes where * can be replaced by BERT, XLNet,
  and GPT2
- Fix pylint & mypy errors
