+++ .pylintrc
@@ -68,6 +68,7 @@ confidence=
 disable=invalid-name,too-many-branches,too-many-statements,too-many-arguments,  # Unnecessarily strict checks
         too-many-instance-attributes,too-few-public-methods,too-many-locals,
         too-many-lines,too-many-return-statements,too-many-boolean-expressions,
+        too-many-ancestors,
         fixme,
         no-else-return,no-else-raise,len-as-condition,unnecessary-pass,         # Not exactly good conventions
         bad-continuation,inconsistent-return-statements,stop-iteration-return,
