+++ texar/modules/classifiers/bert_classifier.py
@@ -17,8 +17,8 @@ BERT classifiers.
 from typing import Optional, Tuple
 
 import torch
-import torch.nn as nn
-import torch.nn.functional as F
+from torch import nn
+from torch.nn import functional as F
 
 from texar.core.layers import get_initializer
 from texar.hyperparams import HParams
