+++ texar/modules/classifiers/gpt2_classifier.py
@@ -17,8 +17,8 @@ GPT2 classifiers.
 from typing import Optional, Tuple
 
 import torch
-import torch.nn as nn
-import torch.nn.functional as F
+from torch import nn
+from torch.nn import functional as F
 
 from texar.core.layers import get_initializer
 from texar.hyperparams import HParams
