+++ texar/modules/classifiers/xlnet_classifier.py
@@ -18,8 +18,8 @@ XLNet Classifier.
 from typing import Any, Dict, Optional, Tuple
 
 import torch
-import torch.nn as nn
-import torch.nn.functional as F
+from torch import nn
+from torch.nn import functional as F
 
 from texar.core.layers import get_initializer
 from texar.hyperparams import HParams
